+++ src/mantle/testing/TestCase.php
@@ -30,6 +30,7 @@ use Mantle\Testing\Concerns\Interacts_With_Hooks;
 use Mantle\Testing\Concerns\Interacts_With_Mail;
 use Mantle\Testing\Concerns\Interacts_With_PHPUnit;
 use Mantle\Testing\Concerns\Interacts_With_Requests;
+use Mantle\Testing\Concerns\Interacts_With_Time;
 use Mantle\Testing\Concerns\Interacts_With_User_Agent;
 use Mantle\Testing\Concerns\Makes_Http_Requests;
 use Mantle\Testing\Concerns\Network_Admin_Screen;
@@ -71,6 +72,7 @@ abstract class TestCase extends BaseTestCase {
 	use Interacts_With_Mail;
 	use Interacts_With_PHPUnit;
 	use Interacts_With_Requests;
+	use Interacts_With_Time;
 	use Interacts_With_User_Agent;
 	use Makes_Http_Requests;
 	use MatchesSnapshots;